% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist_of.R
\name{edgelist_of}
\alias{edgelist_of}
\title{Edgelist_of function}
\usage{
edgelist_of(
  target,
  target_title = F,
  full_URL = T,
  links_title = F,
  origin = F
)
}
\arguments{
\item{target}{The target file you are targeting (Usually a web adress).}

\item{target_title}{Logical value set by default as FALSE. If set to TRUE, it will get the edgelist with the title of the target instead of its url.}

\item{full_URL}{A logical value indicating if you want to return the full URL adresses of the links.}

\item{links_title}{Logical value set by default as FALSE. If set to TRUE, it will return the edgelist with the title of all links. If set to TRUE, it can take a long time for large datasets.}

\item{origin}{A logical value indicating if you want to return the origin of the link (external/internal) in the final dataframe.}
}
\value{
Returns a dataframe object with the information of the edgelist of the connections of the targeted webpage.
}
\description{
This function allows to get an edgelist of the links of a certain webpage. The edge list also includes a column of source, that takes E if external page, and I if it is an internal page.
}
\examples{

target = "https://en.wikipedia.org/wiki/R_(programming_language)"
el <- edgelist_of(target)
summary (el)
}
