% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_from_webpage.R
\name{graph_from_webpage}
\alias{graph_from_webpage}
\title{Graph_from_webpage function : This function allows to get a graph object from a certain webpage.}
\usage{
graph_from_webpage(target)
}
\arguments{
\item{target}{The target file you are targeting (Usually a web adress).}
}
\description{
Graph_from_webpage function : This function allows to get a graph object from a certain webpage.
}
\examples{

target = "https://en.wikipedia.org/wiki/R_(programming_language)"

g <- graph_from_webpage(target)

plot(g,
    layout = igraph::layout_with_fr,
    vertex.size=4,
    vertex.label=NA,
    vertex.label.dist=0.5,
    vertex.color="red",
    edge.arrow.size=0.5)
}
