% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links_of.R
\name{links_of}
\alias{links_of}
\title{Links_of function : This function gets the links of a certain webpage.}
\usage{
links_of(target, full_URL = T, titles = F, origin = F)
}
\arguments{
\item{target}{The target file you are targeting (Usually a web adress).}

\item{full_URL}{A logical value indicating if you want to return the full URL adresses of the links.}

\item{titles}{A logical value indicating if you want to return the titles of the links of the target webpage.}

\item{origin}{A logical value indicating if you want to return the origin of the link (external/internal) in the final dataframe.}
}
\value{
A dataframe with the links of a certain webpage, its source nature (external or internal), and some additional informaiton.
}
\description{
Links_of function : This function gets the links of a certain webpage.
}
\examples{
target = "https://en.wikipedia.org/wiki/R_(programming_language)"
links_of(target)

}
