% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_from_webpage.R
\name{network_from_webpage1}
\alias{network_from_webpage1}
\title{Network_from_webpage1 function : This function allows to get a graph object from a certain webpage. The network_from_webpage function differs from the get_graph function due to the fact that get_graph only gets the graph for one page, and network_from_webpage gets the graph of the connected pages also.}
\usage{
network_from_webpage1(target, iteration = 1)
}
\arguments{
\item{target}{The target file you are targeting (Usually a web adress).}

\item{iteration}{The number of iteration you want. If set to one, this function is equivalent to the "graph_from_webpage() function."}
}
\description{
Network_from_webpage1 function : This function allows to get a graph object from a certain webpage. The network_from_webpage function differs from the get_graph function due to the fact that get_graph only gets the graph for one page, and network_from_webpage gets the graph of the connected pages also.
}
\examples{

target = "https://www.google.com/"

N <- network_from_webpage1(target, iteration = 1)
class(N)
dim(N)

N2 <- network_from_webpage1(target, iteration = 2)
class(N2)
dim(N2)

}
