% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title_of.R
\name{title_of}
\alias{title_of}
\title{Title_of function : This function gets the title of a certain webpage.}
\usage{
title_of(target, collapse = T)
}
\arguments{
\item{target}{The target file you are targeting (Usually a web adress).}

\item{collapse}{A logical that indicates if you want to merge titles in case of the webpage having more than one title element.}
}
\value{
The vector of titles.
}
\description{
Title_of function : This function gets the title of a certain webpage.
}
\examples{
target <- "https://en.wikipedia.org/wiki/R_(programming_language)"
title_of(target)
}
